% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitalCounter.R
\name{DigitCounter_single}
\alias{DigitCounter_single}
\title{Count Numbers of Integer Digits of a Number}
\usage{
DigitCounter_single(data, na.rm = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{data}{A numeric value.}

\item{na.rm}{A logical input indicating whether missing values should be removed before the computation proceeds.}

\item{verbose}{A logical input indicating whether keeps the user up-to-date with messages that print to screen as the function is running.}

\item{...}{Other arguments passed on to methods.}
}
\value{
A value indicating the numbers of integer digits of input.
}
\description{
Count Numbers of Integer Digits of a Number
}
\examples{
DigitCounter_single(43458.99, verbose=TRUE)
DigitCounter_single(0.99, verbose=TRUE)
}
